+++ CHANGELOG.md
@@ -11,6 +11,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/)
 - Peer events: Peer online/offline events are deprecated and will be removed in a future release. Please use channel status events instead. (@Primexz)
 
 ### Fixed
+- Fixed issue where chain sync lost event was getting triggered incorrectly. (@Primexz)
+
 ### Changed
 ### Removed